chore: move non-layout components out of layout folder
