+++ src/components/CategoryItem.jsx
